Moved app components to their own folder and andd propTypes to components.
