+++ package.json
@@ -4,10 +4,11 @@
   "private": true,
   "dependencies": {
     "dotenv": "^16.0.0",
+    "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.0"
+    "react-scripts": "^2.1.8"
   },
   "scripts": {
     "start": "react-scripts start",
