+++ src/components/InputWithLabel.js
@@ -1,5 +1,5 @@
 import React, { useRef , useEffect } from 'react';
-
+import propTypes from 'prop-types';
 const InputWithLabel= ({children, toDoTitle, handleTitleChange}) => {
     const inputRef= React.useRef();
   useEffect(()=>
@@ -10,6 +10,12 @@ const InputWithLabel= ({children, toDoTitle, handleTitleChange}) => {
             <input id="toDoTitle" type="text" name="title" value={toDoTitle} onChange={handleTitleChange} ref={inputRef}/>
         </>
     ) 
+}
+
+InputWithLabel.propTypes = {
+    children: propTypes.string,
+    toDoTitle: propTypes.string.isRequired,
+    handleTitleChange: propTypes.func.isRequired
 }
 
 export default InputWithLabel;
