+++ src/components/ToDoList.module.css
@@ -30,7 +30,6 @@
 .ListItem {
     margin: 5px;
     padding: 5px;
-    width: 80%;
 }
 
 .AddToDoButton {
@@ -38,19 +37,20 @@
    color: rgb(54,73,78);
    font-size: medium;
    width: 4em;
+   margin: 0px 5px 0px 5px;
    border: 'none';
    border-radius: 12px;
 }
 
 .InputBar {
-    margin: 4px;
+    margin: 2px 6px 0px 6px;
+    padding: 2px;
 }
 
 .RemoveButton {
     border: 'none';
-    background-image: url("./Images/trash.png");
-    width: 25px;
-    height: 25px;
+    width: 20px;
+    height: 20px;
     margin: 0px 5px 0px 5px;
 }
 
@@ -70,6 +70,5 @@ ul {
 
 input {
     background-color: rgb(230,230,234);
-    margin: 2px 5px 0px 2px;
     border-radius: 4px;
 }
